requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
black==22.12.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.0.8
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==23.5
django-jet-reboot==1.3.5
django-model-utils==4.3.1
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gopay==2.0.1
html5lib==1.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
mypy-extensions==1.0.0
oauthlib==3.2.2
oscrypto==1.3.0
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.3.3
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.18
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.4
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
reportlab==4.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==5.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.16.0
xhtml2pdf==0.2.15
